8.2 - Implement a test example and an unreliable DB query service.
The purpose of using an unreliable DB query service in local environment
is to unsure the good behavior of the application even if the DB service
fail very often (1 out of 2 queries fails). This may happen in microservice
architecture; the 'retry' helper function is heavy used in the
LinkController for this purpose.